requirements.txt
appdirs==1.4.3
beautifulsoup4==4.8.0
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
httplib2==0.13.1
idna==2.8
oauthlib==3.1.0
pkg-resources==0.0.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
requests==2.22.0
requests-oauthlib==1.2.0
robotframework==3.1.2
robotframework-seleniumlibrary==3.3.1
rsa==4.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.3
tqdm==4.35.0
uritemplate==3.0.0
urllib3==1.25.3
webdrivermanager==0.7.4
